pyproject.toml
[tool.poetry]
name = "daily-upd"
version = "0.1.0"
description = ""
authors = ["seiilaa <vseiilaa@gmail.com>"]
readme = "README.md"
packages = [{include = "daily_upd"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
colorama = "^0.4.6"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
